Revert one gui.cpp change

+++ src/editor/Gui.cpp
@@ -82,7 +82,7 @@ void Gui::init()
 	};
 
 	ifd::FileDialog::Instance().DeleteTexture = [](void* tex){
-		GLuint texID = (GLuint)(tex);
+		GLuint texID = (GLuint)((uintptr_t)tex);
 		glDeleteTextures(1, &texID);
 	};
